+++ src/components/code-editor/code-mirror.jsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import CodeMirrorComp from "@uiw/react-codemirror";
 import { xcodeLight, xcodeDark } from "@uiw/codemirror-theme-xcode";
 import { javascript } from "@codemirror/lang-javascript";
-import EditorContext from "../provider";
+import EditorContext from "./provider";
 
 const CodeMirror = () => {
   const { darktheme, code, setCode } = React.useContext(EditorContext);
