+++ src/components/code-editor/error-bar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import EditorContext from "../provider";
+import EditorContext from "./provider";
 
 const errStyle = {
   color: "red",
