+++ src/components/code-editor/index.css
