+++ src/components/code-editor/index.stories.jsx
