+++ src/components/code-editor/text-editor.jsx
@@ -1,5 +1,5 @@
 import React, { useRef } from 'react'
-import EditorContext from "../provider";
+import EditorContext from "./provider";
 
 const TextEditor = () => {
   const { code } = React.useContext(EditorContext);
